backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
asgiref
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv